Move components used by MRVA out of remote-queries
