+++ extensions/ql-vscode/src/stories/data/remoteQueryResultMessage.json
