+++ extensions/ql-vscode/src/stories/remote-queries/RemoteQueries.stories.tsx
@@ -5,8 +5,8 @@ import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import { RemoteQueries } from "../../view/remote-queries/RemoteQueries";
 
-import * as remoteQueryResult from "./data/remoteQueryResultMessage.json";
-import * as analysesResults from "./data/analysesResultsMessage.json";
+import * as remoteQueryResult from "../data/remoteQueryResultMessage.json";
+import * as analysesResults from "../data/analysesResultsMessage.json";
 
 export default {
   title: "MRVA/Remote Queries",
