+++ extensions/ql-vscode/src/view/variant-analysis/AnalysisAlertResult.tsx
